Rename default helper of icmaa-category
