+++ src/modules/icmaa-category/components/List.ts
@@ -1,6 +1,6 @@
 import { mapGetters } from 'vuex'
 import { CategoryStateCategory } from '../types/CategoryState'
-import { extractPrefix } from '../helpers/fetchCategories'
+import { extractPrefix } from '../helpers'
 import { htmlDecode } from '@vue-storefront/core/lib/store/filters';
 
 interface Letter {
